backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
pandas
numpy
pymongo
python-dotenv
reportlab
pdfplumber
ollama
langgraph
langchain